+++ src/lib/deploy.js
@@ -11,6 +11,7 @@ import util from 'util';
 import step from './helpers.js';
 
 const log = console.log;
+const DEFAULT_NETWORK_ID = 'testnet';
 const DEFAULT_GRAPHQL = 'https://proxy.berkeley.minaexplorer.com/graphql'; // The endpoint used to interact with the network
 
 /**
@@ -221,6 +222,10 @@ export async function deploy({ alias, yes }) {
   }
   let { PrivateKey, Mina, AccountUpdate } = await import(o1jsImportPath);
 
+  // We need to default to the testnet networkId if none is specified for this deploy alias in config.json
+  // This is to ensure the backward compatibility.
+  const networkId =
+    config.deployAliases[alias]?.networkId ?? DEFAULT_NETWORK_ID;
   const graphQlUrl = config.deployAliases[alias]?.url ?? DEFAULT_GRAPHQL;
 
   const { data: nodeStatus } = await sendGraphQL(
@@ -389,7 +394,10 @@ export async function deploy({ alias, yes }) {
   }
 
   let transaction = await step('Build transaction', async () => {
-    let Network = Mina.Network(graphQlUrl);
+    const Network = Mina.Network({
+      networkId,
+      mina: graphQlUrl,
+    });
     Mina.setActiveInstance(Network);
     let tx = await Mina.transaction({ sender: feepayerAddress, fee }, () => {
       AccountUpdate.fundNewAccount(feepayerAddress);