+++ tests/cli/config.spec.ts
@@ -1,8 +1,10 @@
 import { expect, test } from '@playwright/test';
 import { prepareEnvironment } from '@shimkiv/cli-testing-library';
+import { NetworkId } from 'mina-signer';
 import crypto from 'node:crypto';
 import fs from 'node:fs';
 import Constants from '../../src/lib/constants.js';
+import { generateInputsForOptionSelection } from '../utils/cli-utils.js';
 import {
   TestConstants,
   cleanupFeePayerCache,
@@ -46,6 +48,7 @@ test.describe('zkApp-CLI', () => {
 
   test(`should not create deployment alias if procedure was cancelled, @parallel @smoke @config @fail-cases`, async () => {
     const projectName = crypto.randomUUID();
+    const networkId: NetworkId = 'testnet';
     const deploymentAlias = crypto.randomUUID();
     const feePayerAlias = crypto.randomUUID();
     const feePayerAccount = await acquireAvailableAccount();
@@ -60,6 +63,8 @@ test.describe('zkApp-CLI', () => {
       await test.step('Deployment alias creation cancellation', async () => {
         const { exitCode } = await zkConfig({
           processHandler: spawn,
+          workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias,
           feePayerAlias,
           feePayerAccount,
@@ -92,6 +97,7 @@ test.describe('zkApp-CLI', () => {
 
   test(`should properly validate input field values, @serial @smoke @config`, async () => {
     const projectName = crypto.randomUUID();
+    const networkId: NetworkId = 'testnet';
     const deploymentAlias = crypto.randomUUID();
     const feePayerAlias = crypto.randomUUID();
     // Serial because of  the following new aliases
@@ -109,6 +115,8 @@ test.describe('zkApp-CLI', () => {
         await zkProject(projectName, 'none', true, spawn);
         await zkConfig({
           processHandler: spawn,
+          workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias,
           feePayerAlias,
           feePayerAccount,
@@ -158,6 +166,10 @@ test.describe('zkApp-CLI', () => {
               newDeploymentAlias,
               'enter',
             ],
+            'Choose the target network': generateInputsForOptionSelection(
+              networkId,
+              Constants.networkIds
+            ),
             'Set the Mina GraphQL API URL to deploy to': ['enter'],
             // TODO: Add more URL validation cases after the fix of:
             // - https://github.com/o1-labs/zkapp-cli/issues/428
@@ -216,6 +228,7 @@ test.describe('zkApp-CLI', () => {
 
         checkZkConfig({
           workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias: newDeploymentAlias,
           feePayerAlias: newFeePayerAlias,
           feePayerAccount,
@@ -236,6 +249,7 @@ test.describe('zkApp-CLI', () => {
 
   test(`should create deployment alias in case of no cached fee payer account available, @serial @smoke @config`, async () => {
     const projectName = crypto.randomUUID();
+    const networkId: NetworkId = 'testnet';
     const deploymentAlias = crypto.randomUUID();
     const feePayerAlias = crypto.randomUUID();
     const feePayerAccount = await acquireAvailableAccount();
@@ -255,6 +269,8 @@ test.describe('zkApp-CLI', () => {
       await test.step('Deployment alias creation and results validation', async () => {
         const { exitCode, stdOut } = await zkConfig({
           processHandler: spawn,
+          workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias,
           feePayerAlias,
           feePayerAccount,
@@ -267,6 +283,7 @@ test.describe('zkApp-CLI', () => {
         });
         checkZkConfig({
           workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias,
           feePayerAlias,
           feePayerAccount,
@@ -286,6 +303,7 @@ test.describe('zkApp-CLI', () => {
 
   test(`should create deployment aliases using available fee payer account management approaches, @serial @smoke @config`, async () => {
     const projectName = crypto.randomUUID();
+    const networkId: NetworkId = 'testnet';
     let deploymentAlias = crypto.randomUUID();
     let feePayerAlias = crypto.randomUUID();
     const feePayerAccount = await acquireAvailableAccount();
@@ -309,6 +327,8 @@ test.describe('zkApp-CLI', () => {
         await test.step(`Deployment alias creation (feePayerMgmtType=${feePayerMgmtType}) and results validation`, async () => {
           const { exitCode, stdOut } = await zkConfig({
             processHandler: spawn,
+            workDir: `${path}/${projectName}`,
+            networkId,
             deploymentAlias,
             feePayerAlias,
             feePayerAccount,
@@ -321,6 +341,7 @@ test.describe('zkApp-CLI', () => {
           });
           checkZkConfig({
             workDir: `${path}/${projectName}`,
+            networkId,
             deploymentAlias,
             feePayerAlias,
             feePayerAccount,
@@ -336,6 +357,8 @@ test.describe('zkApp-CLI', () => {
       await test.step(`Deployment alias creation (feePayerMgmtType=another) and results validation`, async () => {
         const { exitCode, stdOut } = await zkConfig({
           processHandler: spawn,
+          workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias,
           feePayerAlias,
           feePayerAccount,
@@ -348,6 +371,7 @@ test.describe('zkApp-CLI', () => {
         });
         checkZkConfig({
           workDir: `${path}/${projectName}`,
+          networkId,
           deploymentAlias,
           feePayerAlias,
           feePayerAccount,
@@ -364,4 +388,6 @@ test.describe('zkApp-CLI', () => {
       await cleanup();
     }
   });
+
+  // TODO: Respect the config.networkId property (check that saved data reflects user's selection).
 });