+++ CHANGELOG.md
@@ -21,7 +21,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
-- Possibility to configure the target network ID during the `zk config` process. [#564](https://github.com/o1-labs/zkapp-cli/pull/564)
+- Possibility to configure the target network to testnet or mainnet during the `zk config` process. [#564](https://github.com/o1-labs/zkapp-cli/pull/564)
 
 ## [0.16.0](https://github.com/o1-labs/zkapp-cli/compare/v15.2...v16.0) - 2023-12-11